pyproject.toml
[project]
name = "invoicescout"
version = "0.1.0"
description = "InvoiceScout - scan Drive invoices, extract data with OpenRouter, and write to Sheets"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.95.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
scan-invoices = "invoice_scanner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["invoice_scanner"]

[dependency-groups]
dev = [
    "pytest-cov>=5.0.0",
    "pytest>=9.0.2",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501", # Line too long (handled by formatter)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
